Implementing sftp-access fethcer key missing error treatment

+++ scripts/sftp-access.sh
@@ -8,6 +8,12 @@ LOCAL_RESULTS_DIR=$5
 IMAGE_NAME=$6
 PRIVATE_KEY_PATH=/tmp/keys/fetcher_key_rsa
 
+if [ -f $PRIVATE_KEY_PATH ]
+then
+  echo "Missing fetcher private key"
+  exit 1
+fi
+
 #TODO: add log
 echo -e "get -r $REMOTE_RESULTS_DIR\nquit\n" | sftp -i $PRIVATE_KEY_PATH -o "StrictHostKeyChecking=no" -P $FTP_SERVER_PORT $FTP_SERVER_USER@$FTP_SERVER_IP
 mv $IMAGE_NAME/* $LOCAL_RESULTS_DIR
